Update client/src/Billing/CheckoutForm/CheckoutForm.js
The CardElement component imported from react-stripe-elements creates a “card” type element that mounts on the page when the component renders. The CardElement includes inputs for all of the major card fields: the card number, the expiration date, and the CVC. To display those inputs separately, you can use other Element components provided by the library.

The CheckoutForm class defines a component that displays a CardElement and a button for completing the purchase. The button’s click event is wired to the submit method, which is left blank for now—the next step of this tutorial demonstrates how to make the submit method tokenize card data and send it to the server.

The injectStripe function wraps the component, creating a new component with an injected stripe prop, which contains a Stripe object. You must use the wrapped component in your application instead of the original CheckoutForm.