Update client/src/Billing/Billing.js
StripeProvider initializes Stripe and passes in your publishable key. It’s equivalent to creating a Stripe instance with Stripe.js.

The Elements component, which encloses the wrapped checkout form, creates an Elements group. When you use multiple Elements components instead of the combined CardElement, the Elements group indicates which ones are related. For example, if you used separate components for the card number, expiration date, and CVC, you would put them all in the same Elements group. Note that Elements must contain the component that you wrapped with injectStripe, you cannot put Elements inside of the component that you wrap with injectStripe.