Update client/src/Billing/CheckoutForm/CheckoutForm.js
At this point, the application is capable of displaying a card entry form. The next step is adding functionality to the CheckoutForm component’s submit method so that clicking the Purchase button tokenizes the card information and sends it to the server.

In the submit method, tokenize the card information by invoking createToken on the stripe prop. The stripe prop is available inside the component due to the use of injectStripe.